requirements.txt
Pillow==5.4.1
six==1.12.0
torch==1.0.1.post2
torchvision==0.2.2.post3
tornado==6.0.1
uvloop==0.12.1